Partial fix working towards getting rid of "libs"; see #1492.
removed libserial from the tools set
added libserial to the list of dev tools in the docs
added libserial to the list of debian dependencies
fixed the gLiveSupport makefile to use pkg-config for libserial
some whitespace changes

Campcaster compiles and runs after the change.
I have not tested package creation, but I don't expect any issues with it.

This change will make it more difficult to install Campcaster on Gutsy and earlier, but I don't think anybody cares about that.