CC-1713 - New storage directory structure Changed the GUNID to a 32 hex string (MD5 value), there is no more conversion between hex strings and bigints in the database anymore.
Added added the file extension to the file name.

Stored the complete path to the file in the database.  This means that
linking doesnt need to create any files at all. (It used to create a
symlink to the file you were importing)

The structure and file naming should be fine - there are 4096 possibilities
for the first directory level, and even will a million files this is only
244 files per sub-directory.  The GUID is fine for the file name