Make git optional during build
Currently the installer fails when using a github tarball that comes without a .git folder (https://github.com/LibreTime/libretime/issues/47#issuecomment-285365745).

With this change build.sh only executes git if a .git folder is detected. It uses the name of the downloaded folder as part of the version string if a tarball is used.

I plan on releasing travis-ci built tarballs that contain the tagged version (once we have one) and also have a pre-downloaded composer vendor folder. Those will already contain a VERSION file as well and this change already contains the setup needed for that.