Fix release.sh script
Make it add vendor to the tarball and just be a generally nicer script with less verbose output and no abaility to wipe my working copies .git folder when executed in the wrong dir.

I plan on using this to create tarballs that can be used for the rpmbuild process without the build box needing to run composer and connect to the internet. The openSUSE Build Service boxen do not have a connection to the internet during builds.