Restart services always with systemd
This essentially results in the same behaviour legacy upstream was exploiting in upstart. Normally I would argue that depending on such a feature as part of an applications runtime feature-set is bad. This applies, but until we can get that sorted at another leevel this makes everything work as intendend when running tthrough systemd.