Disable media-monitor on update
The installer originally contained code to stop media-monitor on
updates. This code was removed together with media-monitor. In
the update case we want to make sure that media-monitor is never
running after updating to LibreTime.

The stop command will fail with a warning on systems that never
had media-monitor installed. This warning can be ignored and the
other services stopped with the same command still get stopped
as intended.

Fix #230