fix: update preview_jplayer.js
Added support for previewing FLAC files in the web interface, provided in #509 by marmotte32 on Github in the comments for this issue.

I have tested this against a script-installed copy of 4.2.0, and FLAC preview is working, although auto-play isn't.  I haven't tested to confirm if this behavior matches MP3 and OGG uploads.