fix(legacy): now macro should use station timezone (#3143)
### Description

The `now{}` macro for criteria filtering used UTC rather than station
timezone leading to confusing results.

**This is a new feature**:

No

**I have updated the documentation to reflect these changes**:

Yes

### Testing Notes

**What I did:**

I named a track the current date and hour, and used the date macro to
try to select it. It failed, I then implemented this fix and saw that
the track was selected.

**How you can replicate my testing:**

See what I did.

### **Links**

None