fix(legacy): scheduled show length should not add track fade times (#3144)
### Description

When generating a show that contains a playlist with dynamic blocks, the
show will be incorrectly under scheduled because the fade time was added
onto the total track length. Fade should not add time to an overall
track length as then it wouldn't be a fade.

**This is a new feature**:

No

**I have updated the documentation to reflect these changes**:

No, none needed.

### Testing Notes

**What I did:**

I created a playlist with 3 dynamic blocks:

1. 10 minute length dynamic block
2. 1 item dynamic block
3. fill rest of show dynamic block

I then created a show with the defined auto loading playlist. I then saw
that on auto-load the show was not under scheduled.

**How you can replicate my testing:**

See testing notes

Co-authored-by: Kyle Robbertze <paddatrapper@users.noreply.github.com>