feat: build and run custom nginx container (#3155)
### Description

This replaces the volume mount that shared PHP code between the nginx
and legacy containers with a custom nginx image that has the code baked
in from build. I would have preferred to find a solution that did not
require this, but it looks like docker volumes do not support temporary
volumes that only last the lifetime of the container.

**This is a new feature**:

No

**I have updated the documentation to reflect these changes**:

N/A

### Testing Notes

**What I did:**

Tested with docker-compose

**How you can replicate my testing:**

Run the docker-compose setup

### **Links**

Closes: #3150

---------

Co-authored-by: jo <ljonas@riseup.net>